Fixes to database structure and the installer.
git-svn-id: svn://suivarro/touchstone/trunk@4150 8d9c46d6-3bce-41d5-b1ab-56de0fb72b2f

+++ install/staff_help.sql
@@ -29,7 +29,7 @@ CREATE TABLE `staff_help` (
   `body_plain` text,
   `type` enum('page','pointer') DEFAULT NULL,
   `checkout_time` datetime DEFAULT NULL,
-  `checkout_authorID` mediumint(8) unsigned DEFAULT NULL,
+  `checkout_authorID` int(10) unsigned DEFAULT NULL,
   `roles` enum('SysAdmin','Admin','Staff') DEFAULT NULL,
   `deleted` datetime DEFAULT NULL,
   PRIMARY KEY (`id`),
